add support for STAR-Fusion import to spreadsheet view
